refactor: move qq_chat_agent steer handling into qq_chat_agent_steer.rs
Centralize all steer-related logic into a dedicated module:
- QqChatSteerHook and BrainIterationHook implementation
- build_steer_user_message / build_merged_steer_user_message helpers
- Session-busy steer enqueue and command bypass
- Steer follow-up loop for claimed sessions

This reduces scattering across qq_chat_agent.rs, qq_chat_agent_core.rs,
and qq_chat_agent_claimed.rs and improves maintainability.
